requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
protobuf==4.22.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dotenv==1.0.0
requests==2.30.0
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==2.0.2